feat: Add Debt entity to dashboard module
This commit adds the Debt entity to the dashboard module. The Debt entity is imported and included in the TypeOrmModule.forFeature() method in the dashboard.module.ts file. Additionally, the Debt entity is imported in the dashboard.service.ts file and added to the InjectRepository decorator. This allows the dashboard service to fetch and display debt data along with expenses and incomes in the dashboard.

Refactor the getDashboardData method in dashboard.service.ts

This commit refactors the getDashboardData method in the dashboard.service.ts file. The method now retrieves the latest debt data for the user, selects the necessary columns, and orders the data by creation date. The debt data is then combined with expense and income data to create a list of recent movements. The total balance is also calculated based on the total expenses and total incomes.

Update DTOs for creating expenses, incomes, and debts

This commit updates the create-expense.dto.ts, create-income.dto.ts, and create-debt.dto.ts files. The creation_date property is added to each DTO as an optional field. This allows users to specify a custom creation date when creating expenses, incomes, and debts.

Update expense and income entities

This commit updates the expense.entity.ts and income.entity.ts files. The spendAt property is renamed to creation_date to better reflect its purpose. The CreateSpendAt method is also updated to set the creation_date property instead of the spendAt property.